+++ src/wrapper/blink/wrap.rs
@@ -1,0 +1,26 @@
+use core::fmt::Debug;
+
+use embedded_graphics::draw_target::DrawTarget;
+use embedded_graphics::iterator::raw::RawDataSlice;
+use embedded_graphics::pixelcolor::raw::BigEndian;
+use embedded_graphics::pixelcolor::{PixelColor, Rgb888};
+use embedded_graphics::text::renderer::TextRenderer;
+use mplusfonts::color::{Invert, Screen, WeightedAvg};
+use mplusfonts::style::BitmapFontStyle;
+
+use crate::backend::DumoBackend;
+#[cfg(feature = "alloc")]
+use crate::wrapper::blink::BlinkWrapper;
+use crate::wrapper::flush::FlushWrapper;
+
+impl<'a, 'b, D, C> BlinkWrapper<DumoBackend<'a, 'b, '_, '_, D, C>, D>
+where
+    C: PixelColor + From<C::Raw>,
+    D: DrawTarget,
+    D::Color: PixelColor + Default + Invert + Screen + WeightedAvg + From<Rgb888>,
+    D::Error: Debug,
+    RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
+    BitmapFontStyle<'a, 'b, D::Color, C, 1>: TextRenderer<Color = D::Color>,
+{
+    impl_with_flush!();
+}