+++ src/wrapper/flush.rs
@@ -1,3 +1,5 @@
+mod wrap;
+
 use core::fmt::Debug;
 use core::marker::PhantomData;