+++ src/wrapper/wrap.rs
@@ -9,7 +9,34 @@ use mplusfonts::color::{Invert, Screen, WeightedAvg};
 use mplusfonts::style::BitmapFontStyle;
 
 use crate::backend::DumoBackend;
+#[cfg(feature = "alloc")]
+use crate::wrapper::blink::BlinkWrapper;
 use crate::wrapper::flush::FlushWrapper;
+
+#[cfg(feature = "alloc")]
+macro_rules! impl_with_blink {
+    () => {
+        /// Returns the backend with a new wrapper around it, redrawing cells to show and hide text
+        /// that should blink. Every time a [`Terminal`] calls the [`Backend::draw`] method as part
+        /// of the rendering process, the wrapper advances the blinking animation by one frame. The
+        /// `long_period` and `short_period` parameters apply to [`SLOW_BLINK`] and [`RAPID_BLINK`]
+        /// modifiers, respectively, where the number of frames in an animation cycle is specified.
+        ///
+        /// Backends without this wrapper display text that is set to blink as solid text.
+        ///
+        /// [`Backend::draw`]: ratatui_core::backend::Backend::draw
+        /// [`Terminal`]: ratatui_core::terminal::Terminal
+        /// [`SLOW_BLINK`]: ratatui_core::style::Modifier::SLOW_BLINK
+        /// [`RAPID_BLINK`]: ratatui_core::style::Modifier::RAPID_BLINK
+        pub const fn with_blink(
+            self,
+            long_period: usize,
+            short_period: usize,
+        ) -> BlinkWrapper<Self, D> {
+            BlinkWrapper::new(self, long_period, short_period)
+        }
+    };
+}
 
 macro_rules! impl_with_flush {
     () => {
@@ -41,5 +68,7 @@ where
     RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
     BitmapFontStyle<'a, 'b, D::Color, C, 1>: TextRenderer<Color = D::Color>,
 {
+    #[cfg(feature = "alloc")]
+    impl_with_blink!();
     impl_with_flush!();
 }